Fixed some compiler warnings on %lli and uint64_t or tint data types.

+++ bin64.cpp
@@ -75,6 +75,6 @@ const char* bin64_t::str () const {
     else if (v==NONE)
         return "(NONE)";
     else
-        sprintf(_b64sr[_rsc],"(%i,%lli)",(int)layer(),offset());
+        sprintf(_b64sr[_rsc],"(%i,%llu)",(int)layer(),(unsigned long long int)offset());
     return _b64sr[_rsc];
 }
